Fix swallowed issues (#40)
In texts with different issues it could happen, that some issues were not found. This happened, because the issues were evaluated after one another. When an issue a was processed after another one b that is mentioned in the text after a, a would not have been found.

This fixes this issue by first searching for all issues in the text and then building the elements after sorting all found issues.